Oops, yesterday's "fix" for #10466 made it even worse :)
2002-02-20  Michael Natterer  <mitch@gimp.org>

	Oops, yesterday's "fix" for #10466 made it even worse :)

	* app/core/gimpdrawable-transform.c: need the 0.5 offset to
	the pixel's center only for INTERPOLATION_NONE, as the LINEAR
	and CUBIC algorithms already know about their errors.

	* app/tools/gimpperspectivetool.c
	* app/tools/gimpscaletool.c
	* app/tools/gimpsheartool.c: some more s/gint/gdouble/ so the
	tools can detect pointer motions again...

+++ app/tools/gimpperspectivetool.c
@@ -213,7 +213,7 @@ static void
 perspective_tool_motion (GimpTransformTool *transform_tool,
 			 GimpDisplay       *gdisp)
 {
-  gint diff_x, diff_y;
+  gdouble diff_x, diff_y;
 
   diff_x = transform_tool->curx - transform_tool->lastx;
   diff_y = transform_tool->cury - transform_tool->lasty;
